+++ package.json
@@ -5,7 +5,7 @@
   "directories": {
     "test": "test"
   },
-  "main": "TestContext.js",
+  "main": "index.js",
   "scripts": {
     "lint": "tslint --project ./src/tsconfig.json",
     "lint:test": "tslint --project ./test/tsconfig.json",
