+++ src/Context/ITestContextOptions.ts
@@ -1,0 +1,3 @@
+export interface ITestContextOptions {
+    callerRecursionDepth: number;
+}
