+++ src/index.ts
@@ -1,0 +1,2 @@
+export { TestContext } from './Context/TestContext';
+export {ITestContextOptions} from './Context/ITestContextOptions';
